refactor: dry native connectors
